moves the ColorPicker component to a folder named common and adjusts the path of the import in the Homepage editor interface component
